uuid 엉ë어노테이션 적용
