+++ fanmon/src/main/java/com/example/fanmon/board/entity/Artistboard.java
@@ -1,8 +1,6 @@
 package com.example.fanmon.board.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
@@ -13,9 +11,18 @@ import java.util.UUID;
 @Table(name="artistboard")
 public class Artistboard {
     @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "artistboarduuid", updatable = false, nullable = false)
     private UUID artistboarduuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "groupartistuuid", updatable = false, nullable = false)
     private UUID groupartistuuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "artistuuid", updatable = false, nullable = false)
     private UUID artistuuid;
+
     private LocalDateTime postdate;
     private String content;
     private long likecount;
