+++ fanmon/src/main/java/com/example/fanmon/board/entity/Boardnotice.java
@@ -1,8 +1,6 @@
 package com.example.fanmon.board.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
@@ -14,10 +12,20 @@ import java.util.UUID;
 public class Boardnotice {
 
     @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "noticeuuid", updatable = false, nullable = false)
     private UUID noticeuuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "fanboarduuid", updatable = false, nullable = false)
+    private UUID fanboarduuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "managementuuid", updatable = false, nullable = false)
+    private UUID managementuuid;
+
     private String title;
     private LocalDateTime postdate;
     private String content;
-    private UUID fanboarduuid;
-    private UUID managementuuid;
+
 }
