+++ fanmon/src/main/java/com/example/fanmon/board/entity/Fanboard.java
@@ -1,8 +1,6 @@
 package com.example.fanmon.board.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
@@ -13,10 +11,22 @@ import java.util.UUID;
 @Table(name="fanboard")
 public class Fanboard {
     @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "fanboarduuid", updatable = false, nullable = false)
     private UUID fanboarduuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "artistuuid", updatable = false, nullable = false)
     private UUID artistuuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "useruuid", updatable = false, nullable = false)
     private UUID useruuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "groupartistuuid", updatable = false, nullable = false)
     private UUID groupartistuuid;
+
     private String title;
     private LocalDateTime postdate;
     private String content;
