+++ fanmon/src/main/java/com/example/fanmon/board/entity/Fancomment.java
@@ -1,8 +1,6 @@
 package com.example.fanmon.board.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
@@ -13,9 +11,18 @@ import java.util.UUID;
 @Table(name="fancomment")
 public class Fancomment {
     @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "fancommentuuid", updatable = false, nullable = false)
     private UUID fancommentuuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "fanboarduuid", updatable = false, nullable = false)
     private UUID fanboarduuid;
+
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "useruuid", updatable = false, nullable = false)
     private UUID useruuid;
+
     private LocalDateTime postdate;
     private String content;
     private long ref;
